messenger: fix PeerConnected empty features bug
If we need to connect directly to the introduction node provided by an offer,
LNDK's onion messenger will receive a PeerConnected event to process this new
connection. When we first look at that new peer, the features map is often empty
and LNDK decides onion support must be "false". This botches future payment
attempts, because the onion messenger thinks that this peer doesn't support
onion messaging, and shouldn't be considered in future payment paths. To fix
this, we add a little wait loop that looks for a non-empty map.