messenger: wait for onion update before sending ir
This fixes a race condition that occurs when we try to make a payment before
a PeerConnected event has reached the onion messenger with the updated
onion support status. To fix this, we add a wait period in the onion messenger
to wait for this update to come in before we send the invoice request.